Move settings_test.py to kitsune/.
