+++ kitsune/settings_test.py
